requirements-fastapi.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
Pillow
torch
transformers
